third_party/blink/web_tests/wpt_internal/hsts/only-top-level-navigation-hsts-upgrade.tentative.sub.html

<!DOCTYPE html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
// Note: this lives in wpt_internal because it's testing a Chromium-specific
// behavior for HSTS Tracking Prevention: specifically that a 3P resource load
// can set HSTS but it's only taken into account for top-level navigations.
// Other browsers have slightly different behaviors:
//  * Firefox does not allow setting HSTS from a 3P context.
//. * Safari only allows HSTS to be set for the currently loaded hostname, or
//    the etld+1

  // Check HSTS upgrades only apply to top-level (outermost) frame navigations.
  // Note that this test must be run on an insecure origin because it relies on
  // insecure iframes being loadable. If it's instead run on a secure origin
  // then mixed content blocking will prevent HSTS from working.

  // 0) Confirm that insecure iframes can be loaded.
  // 1) Pin the alt hostname to the HSTS via hsts.html
  // 2) Attempt to load an iframe via http. This should fail because the WPT
  //    server only accepts HTTPS requests on the specified HTTPS port in
  //    http://{{hosts[alt][]}}:{{ports[https][0]}} *and*
  //    HSTS should not upgrade the iframe navigation to https.
  // 3) Open a new window and navigate it to the same http origin. This should
  //    successfully be upgraded to https, load, and then postMessage its origin.

  const CURRENT_HOST_HTTP_PORT = "{{hosts[][]}}:{{ports[http][0]}}";
  const CURRENT_HOST_HTTPS_PORT = "{{hosts[][]}}:{{ports[https][0]}}";
  const ALTERNATE_HOST_HTTPS_PORT = "{{hosts[alt][]}}:{{ports[https][0]}}";

  promise_test(async() => {

    function onMessageWithTimeout(name) {
      return new Promise((resolve, reject) => {

      const timeoutID = step_timeout(() => {
        reject(new Error("Timeout: Didn't receive message for " + name));
        onmessage = null;
      }, 3000);

      onmessage = (event) => {
        clearTimeout(timeoutID);
        resolve(event);
      };
    });
    };

    // Step 0.
    const iframeLoadable = document.createElement('iframe');
    const iframeLoadablePromise = onMessageWithTimeout("Step 0");

    iframeLoadable.src = `http://${CURRENT_HOST_HTTP_PORT}/wpt_internal/hsts/resources/hsts.html`;
    document.body.appendChild(iframeLoadable);
    await iframeLoadablePromise;

    // Step 1.
    // Add HSTS pin for domain.
    await fetch(`https://${ALTERNATE_HOST_HTTPS_PORT}/wpt_internal/hsts/resources/hsts.html`);

    // Step 2.
    // Note: HTTP, not HTTPS:
    const hstsIframe = document.createElement('iframe');
    const hstsIframePromise = onMessageWithTimeout("Step 2")
    .then(resolve => assert_false(true, "HSTS iframe unexpectedly loaded"),
                                  reject => {/*frame didn't load, as expected */});

    hstsIframe.src = `http://${ALTERNATE_HOST_HTTPS_PORT}/wpt_internal/hsts/resources/hsts.html`;
    document.body.appendChild(hstsIframe);
    await hstsIframePromise;

    // Step 3.
    const hstsWindowPromise = onMessageWithTimeout("Step 3")
    .then((event) =>
      assert_equals(event.data.origin,
                    `https://${ALTERNATE_HOST_HTTPS_PORT}`));

    const w = window.open(`http://${ALTERNATE_HOST_HTTPS_PORT}/wpt_internal/hsts/resources/post-origin-to-opener.html`, "_blank");
    if(!w) {
      assert_false(true, "Window didn't open. Is there a popup blocker?");
    }

    await hstsWindowPromise;
}, "HSTS only navigates top-level");

</script>
</body>
</html>
